'修复前台预约，后台服务类型没有显示，编辑不了bug'

+++ wolfcode-car-business/src/main/java/com/wolfcode/car/appointment/domain/vo/ReservationVo.java
@@ -22,7 +22,7 @@ public class ReservationVo {
     /** 汽车类型 */
     private String carSeries;
     /** 服务类型【维修0/保养1】 */
-    private Integer serviceType;
+    private Long serviceType;
     /** 创建时间 */
     private Date createTime;
     /** 备注信息 */
